Inline _app.tsx into main.tsx
